requirements.txt
Django==1.9.4
djangorestframework==3.3.3
fake-factory==0.5.7
httpie==0.9.3
httplib2==0.9.2
Pygments==2.1.3
python-dateutil==2.5.1
requests==2.9.1
shortuuid==0.4.3
six==1.10.0
